+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -26,6 +26,7 @@ import mozilla.components.concept.engine.utils.EngineVersion
 import mozilla.components.concept.engine.webextension.WebExtension
 import mozilla.components.concept.engine.webextension.WebExtensionDelegate
 import org.json.JSONObject
+import org.mozilla.geckoview.AllowOrDeny
 import org.mozilla.geckoview.ContentBlocking
 import org.mozilla.geckoview.ContentBlockingController
 import org.mozilla.geckoview.ContentBlockingController.Event
@@ -36,6 +37,7 @@ import org.mozilla.geckoview.GeckoSession
 import org.mozilla.geckoview.GeckoWebExecutor
 import org.mozilla.geckoview.WebExtensionController
 import java.lang.IllegalStateException
+import java.util.WeakHashMap
 
 /**
  * Gecko-based implementation of Engine interface.
@@ -158,6 +160,13 @@ class GeckoEngine(
         this.webExtensionDelegate = webExtensionDelegate
 
         val tabsDelegate = object : WebExtensionController.TabDelegate {
+            // We use this map to find the engine session of a given gecko
+            // session, as we currently have no other way of accessing the
+            // list of engine sessions. This will change once the engine
+            // gets access to the browser store:
+            // https://github.com/mozilla-mobile/android-components/issues/4965
+            private val tabs = WeakHashMap<GeckoEngineSession, String>()
+
             override fun onNewTab(
                 webExtension: org.mozilla.geckoview.WebExtension?,
                 url: String?
@@ -166,7 +175,26 @@ class GeckoEngine(
                 val geckoWebExtension = webExtension?.let { GeckoWebExtension(it.id, it.location) }
                 webExtensionDelegate.onNewTab(geckoWebExtension, url ?: "", geckoEngineSession)
 
+                tabs[geckoEngineSession] = webExtension?.id
                 return GeckoResult.fromValue(geckoEngineSession.geckoSession)
+            }
+
+            override fun onCloseTab(
+                webExtension: org.mozilla.geckoview.WebExtension?,
+                session: GeckoSession
+            ): GeckoResult<AllowOrDeny> {
+                val geckoEngineSession = tabs.keys.find { it.geckoSession == session } ?: return GeckoResult.DENY
+
+                return if (webExtension != null && tabs[geckoEngineSession] == webExtension.id) {
+                    val geckoWebExtension = GeckoWebExtension(webExtension.id, webExtension.location)
+                    if (webExtensionDelegate.onCloseTab(geckoWebExtension, geckoEngineSession)) {
+                        GeckoResult.ALLOW
+                    } else {
+                        GeckoResult.DENY
+                    }
+                } else {
+                    GeckoResult.DENY
+                }
             }
         }